+++ src/main/java/subway/application/dto/SectionCreateDto.java
@@ -1,4 +1,4 @@
-package subway.service.dto;
+package subway.application.dto;
 
 public class SectionCreateDto {
